requirements.txt
flask
python-pptx
openai
langchain
requests
beautifulsoup4
PyMuPDF
